Add files via upload
This code is flashed to a Wemos D1 Mini, using a TMC2208 driver to drive a Nema 17 stepper motor. 
The code makes use of Blynk and AccelStepper libraries. 

it simply drives the stepper motor to a set location when a V1 pin is toggled, and has it return to the start position once V1 is toggled back. Code allows for a physical button on the device to toggle V1 value. 

I use this to drive a stepper motor that opens and closes my blinds.